Bumping verison number to 0.6, because of the sheer
volume of additional work.

Most defuns now call perlnow-message to identify
themselves when called if the perlnow-trace variable
is set.  The commands perlnow-tron and perlnow-troff
can be used to set/unset that var-- tron also writes
a marker to the *Messages* buffer, incrementing the
perlnow-counter var to help identify a trace.  There's
also a perlnow-debug var to control additional debugging
messages.

Added experimental var perlnow-recent-pick-global,
a fixed version of perlnow-recent-pick (which itself
is not in wide use).

Renamed perlnow-test-file-given-policy
as      perlnow-test-from-policy

When creating a test from a module, there was an old
feature that tried to identify the name of the nearest
sub, on the theory that that's what you're going to try
to test.  This used to just get pushed into the kill-ring
for lack of anything better to do with it.  Now it's
typically used as part of the default file name to
create, and it and the module name are inserted into the
test file as comments, near point.

There have been improvements to perlnow-sub-name-to-var:
it treats the territory *above* a sub as part of the sub
so that with embedded pod the pod area is included.

There are two new commands perlnow-test-create and
perlnow-test-create-manually, as alternates to create.
Combined with the improvements to perlnow-sub-name-to-var
these work well when run from a module: a small change
that allows a new workflow that looks promising. This
uses an improved default test file naming style, including
the module name and the subname. Some new functions:
  perlnow-new-test-file-name
  perlnow-next-test-prefix

fullauto: this check-in includes work on the new default
test policy, "fullauto", which tries to use the new
default test naming convention.

In perlnow-test-from-policy, commented out the fallback
logic which tries to swap in a recent pick if the
generated name doesn't exist yet, and so on.  It was
interfering with "fullauto" implementation (and I
suspect with other things).  New and or modified functions:
  perlnow-fullauto-test-from-policy
  perlnow-testloc-from-policy
  perlnow-test-file-menu

Updated the standard keymap to run perlnow-test-create and
perlnow-test-create-manually with "a" and "A" (rather
than the old "alt" feature).

Cleaned up the select test menu, path turned into
section label, and *also* path attached to name as a
text-property, so it's read in along with the file.
New and modified functions:
  perlnow-select-forward-hotspot
  perlnow-select-previous-hotspot
  perlnow-select-file
  perlnow-select-file-from-current-line
  perlnow-markup-file-with-path
  perlnow-get-path-from-markedup-name

This is the first check-in of work-in progress
on the "CRS Project", where CRS means "create-from-select"

"CRS Project", where CRS means "create-from-select":
Implementing support for running the test create
features from the select menu buffer has turned out to
be remarkably difficult.  New function:
  perlnow-test-select-menu-p

In this direction, there have been a lot of massaging
and additions to perlnow-open-test-file,
none of which has been completed, and all might be reverted.
This CRS work uses some new functions:
   perlnow-module-from-t-file
   perlnow-extract-hyphenized-from-standard-t-name
   perlnow-bigletter-p

A small bugfix to perlnow-create-with-template
(presumably) has (something or other) to do with this effort.

There's a new function perlnow-metadata-from-buffer
which may (someday) be used more widely, to centralize
a common concern.  At present, it's just used by
perlnow-test-create and perlnow-test-create-manually.

General purpose functions originally intended to choose
best fit test file matches with different criteria
(may not be in use):
  perlnow-filter-list
  perlnow-grep-list
  perlnow-minimum-nonempty-list