Got test 42-perlnow-weird-lib-inside-t-loc.t working, eliminating some long-standing bad behavior with unusual set-ups that have perlnow looking for non-existant "t" directories.
Completely re-vamped argument checking of
  perlnow-set-associated-code-pointers
(You must check in sequence: is it non-nil? is it a string? is it
an existing file?)

perlnow-generate-run-string now returns just
an empty string if not give a program to run.

perlnow-list-test-files, instead of erroring out when given
a non-existing dir, returns an empty list.

And added some more of my manual trace messages.

Fixed some existing tests: keep getting expected and actual confused,
assert-equal has them backwards in my opinion.