perlnow-module (as well as perlnow-object-module) now
have a "harder" feature that creates a cpan-style setup for the
new module.  This uses the same single input approach as the raw
module creation commands, which I have mixed feelings about, but
maybe it's useable.  Note: perlnow-object-module feature was
added in another recent commit, may not have mentioned in commit message.